Rename time.h to wztime.h
