+++ src/multiplay.cpp
@@ -75,7 +75,7 @@
 #include "main.h"								// for gamemode
 #include "multiint.h"
 #include "activity.h"
-#include "time.h"
+#include "wztime.h"
 
 // ////////////////////////////////////////////////////////////////////////////
 // ////////////////////////////////////////////////////////////////////////////
